backend/pyproject.toml
[tool.poetry]
name = "esearch"
version = "0.1.0"
description = ""
authors = ["Guillaume Desforges <guillaume.desforges.pro@gmail.com>"]
readme = "README.md"
include = [ "./esearch" ]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
requests = "^2.31.0"
python-dotenv = "^1.0.0"
langchain = "^0.0.184"
sentence-transformers = "^2.2.2"
pymilvus = "2.2.8"
atlassian-python-api = "^3.38.0"
lxml = "^4.9.2"
rwkv = "^0.7.4"
tokenizers = "^0.13.3"
transformers = "^4.29.2"
beautifulsoup4 = "^4.12.2"
tqdm = "^4.65.0"
sqlalchemy = "^2.0.15"
psycopg2-binary = "^2.9.6"
alembic = "^1.11.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
torch = [
  {version = "^2.0.0", source = "torch-cpu"},
  {version = "^2.0.0", platform = "darwin", source = "pypi" },
]
unstructured = "^0.7.12"
websockets = "^11.0.3"
llama-cpp-python = "^0.1.73"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pyright = "^1.1.309"
httpx = "^0.24.1"
poethepoet = "^0.20.0"
ruff = "^0.0.275"
notebook = "^6.5.4"
matplotlib = "^3.7.1"
mypy = "^1.4.1"
pytest = "^7.4.0"
syrupy = "^4.0.5"
snakeviz = "^2.2.0"


[[tool.poetry.source]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F", "I", "N", "A", "ISC", "PLR", "SIM", "RET", "Q", "ICN", "ARG", "ERA", "C", "ANN", "B", "C4", "G", "PIE", "PYI", "PT", "PD", "PL", "TRY", "NPY", "RUF"]
ignore = ["PLE1205", "G004", "TRY002", "TRY003"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "notebooks"
]

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"


[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.poe.tasks]
al = "poetry run alembic"   
rp = "poetry run python"   
